+++ CHANGES.md
@@ -4,7 +4,7 @@
 
 ### 1.0.2
 * MasterController takes doubleClick action only for TableRow (not TableHeader)
-
+* FXMLoading replaces package path with custom path if configured (see tutorial config)
 
 ### 1.0.1
 * scala 2.13.5